Moved code that assembles the GBA cart into class cart_assembler. This is derived from divided_thumb_assembler, so writing assembly code in there is more convenient, since we do not have to dereference the assembler all the time in order to emit instructions.
Went back to the version of the depacker with no code in the header,
since we're going to change what part of the depacker we're going to stick
into the header.

This version of the depacker produces binaries of size 3128 bytes when
used with lostmarbles.elf from the testdata directory.