Also pass version to find_package
I haven't found this documented, but it appears to work.
If I e.g. specify 1.0.1 to find_package instead and have only 1.0.0
installed, then it apparently doesn't find it and obtains it from git.
If I specify 1.0.0 to find_package and have 1.0.0 installed, then it
apparently finds it, which I can tell from cmake not invoking git.