Remove note on argpppp requiring a callback for options with a key
Of course it does require a callback. You add an option, you handle it.
argpppp could only silently accept the option during parsing and do
nothing (bad), or reject the option when it is encountered with an error,
but that's an error we could have caught earlier by requiring a callback,
so let's leave it as it is. During development of a parser a client
programmer can always temporarily supply a no-op callback until they're
ready to supply the real callback. At least it's their fault if they
forget to do so.